requirements.txt
aiogram==3.13.1
psycopg[binary]==3.2.1
openai==1.51.0
python-dotenv==1.0.1